+++ common/src/main/java/com/github/teamfossilsarcheology/fossil/block/entity/WorktableBlockEntity.java
@@ -96,7 +96,7 @@ public class WorktableBlockEntity extends MachineContainerBlockEntity {
             if (litTime > 0) {
                 dirty = true;
                 fuelStack.shrink(1);
-                if (fuelStack.isEmpty()) {
+                if (fuelStack.isEmpty() && fuelStack.getItem().hasCraftingRemainingItem()) {
                     items.set(WorktableMenu.FUEL_SLOT_ID, new ItemStack(fuelStack.getItem().getCraftingRemainingItem()));
                 }
             }